Avatars selection and smooth rendering
### What's done:
- new endpoint on backend
- avatars selection on frontend
- avatars uploading and rerendering without page reload
