+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/controllers/AvatarController.kt
@@ -1,5 +1,6 @@
 package com.saveourtool.save.backend.controllers
 
+import com.saveourtool.save.authservice.utils.username
 import com.saveourtool.save.backend.ByteBufferFluxResponse
 import com.saveourtool.save.backend.service.OrganizationService
 import com.saveourtool.save.backend.service.UserDetailsService
@@ -19,9 +20,12 @@ import io.swagger.v3.oas.annotations.tags.Tags
 import org.slf4j.Logger
 import org.springframework.http.*
 import org.springframework.http.codec.multipart.FilePart
+import org.springframework.security.access.prepost.PreAuthorize
+import org.springframework.security.core.Authentication
 import org.springframework.web.bind.annotation.*
 import reactor.core.publisher.Mono
 import reactor.kotlin.core.publisher.toMono
+import javax.transaction.NotSupportedException
 
 import kotlin.time.Duration.Companion.days
 import kotlin.time.toJavaDuration
@@ -40,6 +44,36 @@ internal class AvatarController(
     private val organizationService: OrganizationService,
     private val userDetailsService: UserDetailsService,
 ) {
+    @Operation(
+        method = "POST",
+        summary = "Set avatar from existing avatar packages",
+        description = "This endpoint can only be used in case of setting avatar for user from existing resources",
+    )
+    @Parameters(
+        Parameter(
+            name = "resource",
+            `in` = ParameterIn.QUERY,
+            description = "static resource for avatar",
+            required = true
+        ),
+        Parameter(name = "type", `in` = ParameterIn.QUERY, description = "type of avatar", required = true)
+
+    )
+    @GetMapping(path = ["/avatar_update"])
+    fun updateAvatarFromResources(
+        @RequestParam resource: String,
+        @RequestParam type: AvatarType,
+        authentication: Authentication
+    ): Mono<StringResponse> = Mono.just(resource)
+        .map {
+            when (type) {
+                AvatarType.ORGANIZATION -> throw NotSupportedException("Organization upload is not yet supported")
+                AvatarType.USER -> userDetailsService.setAvatarFromResource(authentication.username(), it)
+            }
+        }.map {
+            ResponseEntity.status(HttpStatus.OK).body("Avatar successfully updated with $it")
+        }
+
     /**
      * @param partMono image to be uploaded
      * @param owner owner name
@@ -53,14 +87,25 @@ internal class AvatarController(
         description = "Upload an avatar for user or organization.",
     )
     @Parameters(
-        Parameter(name = "owner", `in` = ParameterIn.QUERY, description = "user name or organization name", required = true),
+        Parameter(
+            name = "owner",
+            `in` = ParameterIn.QUERY,
+            description = "user name or organization name",
+            required = true
+        ),
         Parameter(name = "type", `in` = ParameterIn.QUERY, description = "type of avatar", required = true),
         Parameter(name = FILE_PART_NAME, `in` = ParameterIn.DEFAULT, description = "body of avatar", required = true),
-        Parameter(name = CONTENT_LENGTH_CUSTOM, `in` = ParameterIn.HEADER, description = "size in bytes of avatar", required = true),
+        Parameter(
+            name = CONTENT_LENGTH_CUSTOM,
+            `in` = ParameterIn.HEADER,
+            description = "size in bytes of avatar",
+            required = true
+        ),
     )
     @ApiResponse(responseCode = "200", description = "Avatar uploaded successfully.")
     @ApiResponse(responseCode = "404", description = "User or organization not found.")
     @PostMapping(path = ["/upload"], consumes = [MediaType.MULTIPART_FORM_DATA_VALUE])
+    @PreAuthorize("isAuthenticated()")
     fun uploadImage(
         @RequestPart(FILE_PART_NAME) partMono: Mono<FilePart>,
         @RequestHeader(CONTENT_LENGTH_CUSTOM) contentLength: Long,
@@ -77,16 +122,14 @@ internal class AvatarController(
                 log.info("Saved $contentLength bytes of $avatarKey")
             }
         }
-        .flatMap {
-            blockingToMono {
-                when (type) {
-                    AvatarType.ORGANIZATION -> organizationService.updateAvatarVersion(owner)
-                    AvatarType.USER -> userDetailsService.updateAvatarVersion(owner)
-                }
+        .map {
+            when (type) {
+                AvatarType.ORGANIZATION -> organizationService.updateAvatarVersion(owner)
+                AvatarType.USER -> userDetailsService.updateAvatarVersion(owner)
             }
         }
         .map {
-            ResponseEntity.status(HttpStatus.OK).body("Image was successfully uploaded")
+            ResponseEntity.ok().body(it)
         }
 
     @Operation(
@@ -96,7 +139,12 @@ internal class AvatarController(
     )
     @Parameters(
         Parameter(name = "type", `in` = ParameterIn.PATH, description = "type of avatar", required = true),
-        Parameter(name = "owner", `in` = ParameterIn.PATH, description = "user name or organization name", required = true),
+        Parameter(
+            name = "owner",
+            `in` = ParameterIn.PATH,
+            description = "user name or organization name",
+            required = true
+        ),
     )
     @ApiResponse(responseCode = "200", description = "Returns content of the file.")
     @ApiResponse(responseCode = "404", description = "User not found.")
