+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/service/OrganizationService.kt
@@ -170,14 +170,17 @@ class OrganizationService(
      * @param name
      * @throws NoSuchElementException
      */
-    fun updateAvatarVersion(name: String) {
+    fun updateAvatarVersion(name: String): String {
         val organization = organizationRepository.findByName(name).orNotFound()
         var version = organization.avatar?.substringAfterLast("?")?.toInt() ?: 0
+        val newAvatar = "${AvatarType.ORGANIZATION.toUrlStr(name)}?${++version}"
 
         organization.apply {
-            avatar = "${AvatarType.ORGANIZATION.toUrlStr(name)}?${++version}"
+            avatar = newAvatar
         }.orNotFound { "Organization with name [$name] was not found." }
         organization.let { organizationRepository.save(it) }
+
+        return newAvatar
     }
 
     /**
