+++ save-cloud-common/src/commonMain/kotlin/com/saveourtool/save/utils/Constants.kt
@@ -74,4 +74,7 @@ const val AUTHORIZATION_ROLES = "X-Authorization-Roles"
 @Suppress("NON_EXPORTABLE_TYPE")
 const val DEFAULT_SETUP_SH_TIMEOUT_MILLIS: Long = 60_000L
 
-const val AVATARS_PACKS_DIR: String = "avatar_packs"
+/**
+ * directory with the default avatars package
+ */
+const val AVATARS_PACKS_DIR: String = "/img/avatar_packs"
