+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/contests/ContestGlobalRatingView.kt
@@ -9,6 +9,7 @@ import com.saveourtool.save.entities.ProjectDto
 import com.saveourtool.save.filters.OrganizationFilter
 import com.saveourtool.save.filters.ProjectFilter
 import com.saveourtool.save.frontend.components.basic.AVATAR_ORGANIZATION_PLACEHOLDER
+import com.saveourtool.save.frontend.components.basic.avatarRenderer
 import com.saveourtool.save.frontend.components.basic.table.filters.nameFiltersRow
 import com.saveourtool.save.frontend.components.tables.TableProps
 import com.saveourtool.save.frontend.components.tables.columns
@@ -20,7 +21,6 @@ import com.saveourtool.save.frontend.components.tables.visibleColumnsCount
 import com.saveourtool.save.frontend.components.views.AbstractView
 import com.saveourtool.save.frontend.externals.fontawesome.faTrophy
 import com.saveourtool.save.frontend.utils.*
-import com.saveourtool.save.v1
 
 import js.core.jso
 import react.*
@@ -118,9 +118,7 @@ class ContestGlobalRatingView : AbstractView<ContestGlobalRatingProps, ContestGl
                                 img {
                                     className =
                                             ClassName("avatar avatar-user width-full border color-bg-default rounded-circle")
-                                    src = cellContext.row.original.organization.avatar?.let {
-                                        "/api/$v1/avatar$it"
-                                    } ?: AVATAR_ORGANIZATION_PLACEHOLDER
+                                    src = cellContext.row.original.organization.avatar?.avatarRenderer() ?: AVATAR_ORGANIZATION_PLACEHOLDER
                                     style = jso {
                                         height = 2.rem
                                         width = 2.rem
