+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/usersettings/SettingsViewRightColumn.kt
@@ -5,6 +5,7 @@
 package com.saveourtool.save.frontend.components.views.usersettings
 
 import com.saveourtool.save.frontend.components.views.usersettings.right.*
+import com.saveourtool.save.frontend.components.views.usersettings.right.profile.profileSettingsCard
 import com.saveourtool.save.validation.FrontendRoutes.*
 
 import react.FC
