+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/usersettings/right/profile/Profile.kt
@@ -1,0 +1,162 @@
+/**
+ * rendering for Profile management card
+ */
+
+@file:Suppress("EMPTY_BLOCK_STRUCTURE_ERROR")
+
+package com.saveourtool.save.frontend.components.views.usersettings.right.profile
+
+import com.saveourtool.save.frontend.components.inputform.InputTypes
+import com.saveourtool.save.frontend.components.views.usersettings.*
+import com.saveourtool.save.frontend.components.views.usersettings.right.SettingsInputFields
+import com.saveourtool.save.frontend.components.views.usersettings.right.validation.*
+import com.saveourtool.save.frontend.utils.*
+
+import react.ChildrenBuilder
+import react.FC
+import react.dom.html.ReactHTML.div
+import react.dom.html.ReactHTML.h4
+import react.dom.html.ReactHTML.hr
+import react.dom.html.ReactHTML.textarea
+import react.useState
+import web.cssom.*
+
+const val AVATARS_PACKAGE_COUNT = 9
+
+val profileSettingsCard: FC<SettingsProps> = FC { props ->
+    // === states ===
+    val (settingsInputFields, setSettingsInputFields) = useState(SettingsInputFields())
+    val saveUser = useSaveUser(props, settingsInputFields, setSettingsInputFields)
+
+    // === design ===
+    div {
+        className = ClassName("row px-5 mt-3")
+
+        hr { }
+
+        div {
+            className = ClassName("col-6 text-center")
+            div {
+                className = ClassName("row mb-2")
+                h4 {
+                    +"Add bio and info:"
+                }
+            }
+            div {
+                className = ClassName("row pr-5")
+                div {
+                    className = ClassName("input-group needs-validation")
+                    textarea {
+                        className = ClassName("form-control shadow")
+                        onChange = {
+                            val newSettingsInputFields =
+                                    settingsInputFields.updateValue(InputTypes.FREE_TEXT, it.target.value, "")
+                            setSettingsInputFields(newSettingsInputFields)
+                        }
+                        placeholder = "400 characters"
+                        defaultValue = props.userInfo?.freeText
+                        rows = 10
+                        maxLength = 400
+                    }
+                }
+            }
+        }
+
+        div {
+            className = ClassName("col-6")
+            div {
+                className = ClassName("row mb-2")
+                h4 {
+                    +"Upload or select avatar:"
+                }
+            }
+
+            avatarSelector {
+                this.userInfo = props.userInfo
+                this.userInfoSetter = props.userInfoSetter
+            }
+        }
+    }
+
+    div {
+        className = ClassName("row")
+
+        div {
+            className = ClassName("col mt-2 px-5")
+            extraInformation(props, settingsInputFields, setSettingsInputFields)
+
+            div {
+                className = ClassName("row justify-content-center")
+                buttonBuilder("Save changes", style = "primary", isDisabled = settingsInputFields.containsError()) {
+                    saveUser()
+                }
+            }
+        }
+    }
+}
+
+private fun ChildrenBuilder.extraInformation(
+    props: SettingsProps,
+    settingsInputFields: SettingsInputFields,
+    setSettingsInputFields: FieldsStateSetter
+) {
+    hr { }
+
+    inputForm(
+        props.userInfo?.realName,
+        InputTypes.REAL_NAME,
+        settingsInputFields,
+        setSettingsInputFields,
+        "e.g. John Smith"
+    ) { validateRealName() }
+
+    inputForm(
+        props.userInfo?.company,
+        InputTypes.COMPANY,
+        settingsInputFields,
+        setSettingsInputFields,
+        "e.g. FutureWay Inc."
+    ) { validateCompany() }
+
+    inputForm(
+        props.userInfo?.location,
+        InputTypes.LOCATION,
+        settingsInputFields,
+        setSettingsInputFields,
+        "Beijing, China"
+    ) { validateLocation() }
+
+    inputForm(
+        props.userInfo?.website,
+        InputTypes.WEBSITE,
+        settingsInputFields,
+        setSettingsInputFields,
+        "https://saveourtool.com"
+    ) { validateWebsite() }
+
+    inputForm(
+        props.userInfo?.linkedin,
+        InputTypes.LINKEDIN,
+        settingsInputFields,
+        setSettingsInputFields,
+        "https://linkedin.com/yourname"
+    ) { validateLinkedIn() }
+
+    inputForm(
+        props.userInfo?.gitHub,
+        InputTypes.GITHUB,
+        settingsInputFields,
+        setSettingsInputFields,
+        "https://github.com/yourname"
+    ) { validateGithub() }
+
+    inputForm(
+        props.userInfo?.twitter,
+        InputTypes.TWITTER,
+        settingsInputFields,
+        setSettingsInputFields,
+        "https://x.com/yourname"
+    ) { validateTwitter() }
+
+    hr { }
+}
